+++ src/conversions/geoc.cpp
@@ -33,6 +33,36 @@
 
 PROJ_HEAD(geoc, "Geocentric Latitude");
 
+/*************************************************************************************/
+PJ_COORD pj_geocentric_latitude(const PJ *P, PJ_DIRECTION direction,
+                                PJ_COORD coord) {
+    /**************************************************************************************
+        Convert geographical latitude to geocentric (or the other way round if
+        direction = PJ_INV)
+
+        The conversion involves a call to the tangent function, which goes
+    through the roof at the poles, so very close (the last centimeter) to the
+    poles no conversion takes place and the input latitude is copied directly to
+    the output.
+
+        Fortunately, the geocentric latitude converges to the geographical at
+    the poles, so the difference is negligible.
+
+        For the spherical case, the geographical latitude equals the geocentric,
+    and consequently, the input is copied directly to the output.
+    **************************************************************************************/
+    const double limit = M_HALFPI - 1e-9;
+    PJ_COORD res = coord;
+    if ((coord.lp.phi > limit) || (coord.lp.phi < -limit) || (P->es == 0))
+        return res;
+    if (direction == PJ_FWD)
+        res.lp.phi = atan(P->one_es * tan(coord.lp.phi));
+    else
+        res.lp.phi = atan(P->rone_es * tan(coord.lp.phi));
+
+    return res;
+}
+
 /* Geographical to geocentric */
 static void forward(PJ_COORD &coo, PJ *P) {
     coo = pj_geocentric_latitude(P, PJ_FWD, coo);
