+++ src/ctx.cpp
@@ -213,3 +213,160 @@ PJ_CONTEXT *proj_context_clone(PJ_CONTEXT *ctx) {
 
     return new (std::nothrow) pj_ctx(*ctx);
 }
+
+/*****************************************************************************/
+int proj_errno(const PJ *P) {
+    /******************************************************************************
+        Read an error level from the context of a PJ.
+    ******************************************************************************/
+    return proj_context_errno(pj_get_ctx((PJ *)P));
+}
+
+/*****************************************************************************/
+int proj_context_errno(PJ_CONTEXT *ctx) {
+    /******************************************************************************
+        Read an error directly from a context, without going through a PJ
+        belonging to that context.
+    ******************************************************************************/
+    if (nullptr == ctx)
+        ctx = pj_get_default_ctx();
+    return ctx->last_errno;
+}
+
+/*****************************************************************************/
+int proj_errno_set(const PJ *P, int err) {
+    /******************************************************************************
+        Set context-errno, bubble it up to the thread local errno, return err
+    ******************************************************************************/
+    /* Use proj_errno_reset to explicitly clear the error status */
+    if (0 == err)
+        return 0;
+
+    /* For P==0 err goes to the default context */
+    proj_context_errno_set(pj_get_ctx((PJ *)P), err);
+    errno = err;
+
+    return err;
+}
+
+/*****************************************************************************/
+int proj_errno_restore(const PJ *P, int err) {
+    /******************************************************************************
+        Use proj_errno_restore when the current function succeeds, but the
+        error flag was set on entry, and stored/reset using proj_errno_reset
+        in order to monitor for new errors.
+
+        See usage example under proj_errno_reset ()
+    ******************************************************************************/
+    if (0 == err)
+        return 0;
+    proj_errno_set(P, err);
+    return 0;
+}
+
+/*****************************************************************************/
+int proj_errno_reset(const PJ *P) {
+    /******************************************************************************
+        Clears errno in the context and thread local levels
+        through the low level pj_ctx interface.
+
+        Returns the previous value of the errno, for convenient reset/restore
+        operations:
+
+        int foo (PJ *P) {
+            // errno may be set on entry, but we need to reset it to be able to
+            // check for errors from "do_something_with_P(P)"
+            int last_errno = proj_errno_reset (P);
+
+            // local failure
+            if (0==P)
+                return proj_errno_set (P, 42);
+
+            // call to function that may fail
+            do_something_with_P (P);
+
+            // failure in do_something_with_P? - keep latest error status
+            if (proj_errno(P))
+                return proj_errno (P);
+
+            // success - restore previous error status, return 0
+            return proj_errno_restore (P, last_errno);
+        }
+    ******************************************************************************/
+    int last_errno;
+    last_errno = proj_errno(P);
+
+    proj_context_errno_set(pj_get_ctx((PJ *)P), 0);
+    errno = 0;
+    return last_errno;
+}
+
+/* Create a new context based on the default context */
+PJ_CONTEXT *proj_context_create(void) {
+    return new (std::nothrow) pj_ctx(*pj_get_default_ctx());
+}
+
+PJ_CONTEXT *proj_context_destroy(PJ_CONTEXT *ctx) {
+    if (nullptr == ctx)
+        return nullptr;
+
+    /* Trying to free the default context is a no-op (since it is statically
+     * allocated) */
+    if (pj_get_default_ctx() == ctx)
+        return nullptr;
+
+    delete ctx;
+    return nullptr;
+}
+
+/************************************************************************/
+/*                  proj_context_use_proj4_init_rules()                 */
+/************************************************************************/
+
+void proj_context_use_proj4_init_rules(PJ_CONTEXT *ctx, int enable) {
+    if (ctx == nullptr) {
+        ctx = pj_get_default_ctx();
+    }
+    ctx->use_proj4_init_rules = enable;
+}
+
+/************************************************************************/
+/*                              EQUAL()                                 */
+/************************************************************************/
+
+static int EQUAL(const char *a, const char *b) {
+#ifdef _MSC_VER
+    return _stricmp(a, b) == 0;
+#else
+    return strcasecmp(a, b) == 0;
+#endif
+}
+
+/************************************************************************/
+/*                  proj_context_get_use_proj4_init_rules()             */
+/************************************************************************/
+
+int proj_context_get_use_proj4_init_rules(PJ_CONTEXT *ctx,
+                                          int from_legacy_code_path) {
+    const char *val = getenv("PROJ_USE_PROJ4_INIT_RULES");
+
+    if (ctx == nullptr) {
+        ctx = pj_get_default_ctx();
+    }
+
+    if (val) {
+        if (EQUAL(val, "yes") || EQUAL(val, "on") || EQUAL(val, "true")) {
+            return TRUE;
+        }
+        if (EQUAL(val, "no") || EQUAL(val, "off") || EQUAL(val, "false")) {
+            return FALSE;
+        }
+        pj_log(ctx, PJ_LOG_ERROR,
+               "Invalid value for PROJ_USE_PROJ4_INIT_RULES");
+    }
+
+    if (ctx->use_proj4_init_rules >= 0) {
+        return ctx->use_proj4_init_rules;
+    }
+    return from_legacy_code_path;
+}
